feat: Make the worker stay in https pages
We want to download file in the worker for performance in a near futur.
So for CORS policy compliance, we need the worker on same domain during saveFiles.
The link clicked by the old code was hardcoded in the page and a http link (scrape.com)
We change that to a goto https://... page