Clean template's code
We begin with the cleaning of the template's code to make place for your own.
You can see we're keeping some mandatory functions. They will be overriden as they will be launched by the pilot as the path to follow during konnector's execution.
We're also renaming the Class to fit targeted website.