EnsureNotAuthenticated
In this commit we are starting by ensuring there is no active session on the targeted website so we can begin on a fresh start. As it is (and will be until we create a cozy account and saved some credentials) the first ever konnector's run on the device.
To cover for the automated logout too, we reach the "login successfully" page of the targeted website.
When one or the other element is detected, we're launching `checkAuthenticated` in the worker to determine if the logout button is visible on the page.
If it is not, here we go, we are ready to let the user login ! If it is, we proceed to click on the logout button and wait for the wanted form input element to show up, indicating the logout has been done properly.