EnsureAuthenticated
Now we are absolutely sure there is no active session on the targeted website, we can show the webview to the user and let him login onto the website for us.
For the moment, this is all this function will do, but we will come back on it when we have saved some credentials to use.