Credentials Interception
Here we implement the credentials interception to be able to save them on the devicea little bit later. For this, you need those two worker's functions `onWorkerReady` and `onWorkerEvent`. First we need to tell the pilot we're gonna listen to the worker event, the line is added at the beginning of `ensureAuthenticated` to get the listener ready right at the start.
Then we create a listener on the wanted event. Here we're using the click on the form's submit button. Like this, when the user click on the button, the event is catched and we can scrape the inputs values on the spot, before the page load. If the user used a wrong combination, he will need to click the button again, replacing the first interception by the new.
After that the worker emits an event containing the event name, and its payload (an object containing the credentials). It is catched by the pilot, checked, and the credentials are placed in the store.
This type of interceptions can be done for various other reasons like detecting an error, wait for a DOM modification (like watching a loading spinner finish before taking action)  etc ... it can be very useful.