GetUserDataFromWebsite
In this function, we'll need to find a `sourceAccountIdentifier` to return, so we can create an account and a trigger in the cozy instance. It will also be used to name the destination directory.
If the credentials interception went well, we supposed to have the user's login credential in the store but we're gonna verify this. First check if we found some saved credentials from previous run, so it's already ready for later runs. Then check in the store. We will prefer the intercepted credentials as it is what the user had just typed, he may have changed his login username recently. This way we're ensured to have the most recent credentials every time. Alternatively, the user may sign in with his email, but his profile may show his username or ID instead. If there are no successful interceptions and no saved credentials, we may need to scrape it on the website.
In this case, we meet our first handmade `worker` function !
Named `findValidSAI` (for good mesure, but not mandatory), we're gonna launch it by giving it as the argument to `runInWorker` `pilot's` method.

/!\ It is very important to add every handmade `worker` functions to the `additionalExposedMethodsNames` array at the end of the file, otherwise it will not be "seen" by the `pilot` and will never be executed /!\

In this function we're gonna scrape on the website the needed data. Here for example we're finding the element containing the username, splitting it's content as it is part of a larger sentence and extract "student" as the identifier we want. Note that you may also find an identifier in numerous different ways, it could be intercepted in a response, found in the local/session storage, or hidden in a burger menu so stay creative.