Adapt manifest
First of all you will need to setup the konnector's manifest and adapt it to the targeted website.
Change all needed values where you see fit such as name and slug of the konnector, your name and link in the developer object and all the needed descriptions, category etc ... However, some of the manifest values must not be touched like the version or the permissions.
Think that if you don't really know what to put in replacement, don't touch it, we will adjust if needed when we will review your work.

missed manifest infos

add spaces in manifest name