Fetch
Time to find the sought files and save them in the cozy instance !
Right before the fun begins, we will save the intercepted credentials if any, that way we have them for future execution, and possibly make the login automated so we didn't even need the user to do anything.
Following that, we will navigate to the first page containing the files if needed. We'll find the number of pages we have to scrape (for this example we fixed the limit to 3 pages to avoid scraping the whole 50 of them). With this number, we'll enter a loop to get all the files needed on every pages. Once in the loop, we'll use a handmade `worker`'s function to get and format all the files of the current page. This function will return the formatted files to the pilot for it to store them waiting for the loop to complete. After that, we navigate to the next page and repeat the process until the condition to end the loop is met. Right after, we pass the fullfilled array of files to the `saveFiles` function for it to save them on the cozy instance.