Identity
Now all the fetching has been done, the last thing to do (when possible) is to fetch the user identity. As the website used for this example does not contain any user's account page, we will just show how to build a userIdentity to be saved on a cozy. If needed, navigate to the info page of the user, then just scrap everything usefull for an identity. Names, address, email and phone number are the common data you can find, but depending of the website, you might want to save a social security number or anything else specific from the targeted website. The structure is mandatory, the email must be a string as an array entry, the address and phone must be in an object as an array's entry as well. Sometimes, splitting an address in streets/postcode/city might not be easy, it can only be the "formattedAddress". No keys are specifically mandatory, it will depends on what you can find on the website. But if you find some, the structure is mandatory or it will not be displayed correctly in the cozy instance.
And with that, we're all done, a brand new CliSK can be added to the list !